+++ src/pytest_bdd/compatibility/importlib/resources.py
@@ -1,17 +1,6 @@
-from importlib_resources import *
 from importlib_resources import as_file, files
 
 __all__ = [
-    "Package",
-    "Resource",
-    "ResourceReader",
     "as_file",
-    "contents",
     "files",
-    "is_resource",
-    "open_binary",
-    "open_text",
-    "path",
-    "read_binary",
-    "read_text",
 ]