+++ tests/args/cucumber_expression/test_args.py
@@ -1,4 +1,5 @@
 """StepHandler arguments tests."""
+
 from typing import TYPE_CHECKING
 
 from pytest import mark