+++ tests/feature/test_report.py
@@ -1,4 +1,5 @@
 """Test scenario reporting."""
+
 import re
 from pathlib import Path
 from typing import Optional, Union